Organize python wrappers as a package. Change concept to return native bitarrays instead of Java bitsets
